chore(application-generic): rename the novu integration utils file
