+++ packages/application-generic/src/index.ts
@@ -18,7 +18,7 @@ export * from './utils/email-normalization';
 export * from './utils/digest';
 export * from './utils/object';
 export * from './utils/hmac';
-export * from './utils/novuIntegrations';
+export * from './utils/novu-integrations';
 export * from './decorators/external-api.decorator';
 export * from './decorators/user-session.decorator';
 export * from './tracing';
