+++ packages/application-generic/src/utils/novu-integrations.ts
